% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_problem.R
\name{calc_coef}
\alias{calc_coef}
\title{Calculate Coefficients of P(x)}
\usage{
calc_coef(s_sq, eigen_vec)
}
\arguments{
\item{s_sq}{s-squared, a positive number}

\item{eigen_vec}{a vector}
}
\value{
a vector
}
\description{
Calculate Coefficients of P(x)
}
