% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_problem.R
\name{count_grids}
\alias{count_grids}
\title{Count The r-by-n 0-1 Matrices Without Adjacent 1's in Any Row or Column}
\usage{
count_grids(r = 1, n = 1, vec = TRUE)
}
\arguments{
\item{r}{a positive integer}

\item{n}{a positive integer}

\item{vec}{logical, default value is TRUE}
}
\value{
a number (the n-th number) or a vector (\code{the_re[1:n]})
}
\description{
Count The r-by-n 0-1 Matrices Without Adjacent 1's in Any Row or Column
}
\examples{
(count_grids(r = 1, n = 6))
(count_grids(r = 2, n = 6))
(count_grids(r = 3, n = 6))
(count_grids(r = 4, n = 6))
(count_grids(r = 5, n = 6))
(count_grids(r = 6, n = 6))
(count_grids(r = 7, n = 6))
(count_grids(r = 8, n = 6))
(count_grids(r = 9, n = 6))
}
