% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibos.R
\name{fibo}
\alias{fibo}
\title{Fibonacci Numbers}
\usage{
fibo(i)
}
\arguments{
\item{i}{a non-negative integer}
}
\value{
the ith Fibonacci number
}
\description{
Fibonacci Numbers
}
\examples{
sapply(0:5, fibo)
}
