% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibos.R
\name{fibo_matrix}
\alias{fibo_matrix}
\title{Fibonacci Phi-Matrix Series}
\usage{
fibo_matrix(n = 1)
}
\arguments{
\item{n}{a positive integer}
}
\value{
the n-the Fibonacci Phi-matrix
}
\description{
Fibonacci Phi-Matrix Series
}
\examples{
(Phi_1 <- fibo_matrix(1))
(Phi_2 <- fibo_matrix(2))
(Phi_3 <- fibo_matrix(3))
}
