% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.r
\name{fibonacci}
\alias{fibonacci}
\title{Generate a list of Fibonacci numbers}
\usage{
fibonacci(N)
}
\arguments{
\item{N}{a positive integer or 0}
}
\value{
if N is a integer >= 0, the function returns a list of lenght
N+1 starting a 0 and containing every Fibonacci number till the Nth
else the function returns 'Error'
}
\description{
This function generates a list of Fibonacci numbers starting at 0
and ending with the Nth Fibonacci number.
}
\details{
For this function 0 is regarded as the 0th Fibonacci number
}
\examples{
fibonacci(0)
fibonacci(2)
fibonacci(160)

\dontrun{
fibonacci("a")
fibonacci(-6)
fibonacci(3.5)
}

}
